Common routines for examples
