+++ examples/common/mod.rs
@@ -1,0 +1,6 @@
+//! Common stuff used by the examples
+
+#![allow(dead_code)]
+
+pub mod numeric;
+pub mod rgb_classes;
