+++ examples/extra_trees_classifier.rs
@@ -2,85 +2,21 @@ extern crate forester;
 extern crate image;
 extern crate rand;
 
+mod common;
+
 use std::f64::consts::PI;
 use std::fs::File;
-use std::iter::Sum;
 
 use rand::{thread_rng, Rng};
 
 use forester::array_ops::Partition;
-use forester::categorical::{Categorical, CatCount};
+use forester::categorical::CatCount;
 use forester::data::{SampleDescription, TrainingData};
 use forester::dforest::DeterministicForestBuilder;
 use forester::dtree::DeterministicTreeBuilder;
-use forester::iter_mean::IterMean;
 use forester::split::{BestRandomSplit, Split};
 
-#[derive(Copy, Clone)]
-enum Classes {
-    Red,
-    Green,
-    Blue,
-}
-
-impl Categorical for Classes {
-    fn as_usize(&self) -> usize {
-        *self as usize
-    }
-
-    fn n_categories(&self) -> Option<usize> {
-        Some(3)
-    }
-}
-
-#[derive(Clone)]
-struct ClassCounts {
-    p: [usize; 3],
-}
-
-impl ClassCounts {
-    fn new() -> Self {
-        ClassCounts {
-            p: [0; 3]
-        }
-    }
-}
-
-impl CatCount<Classes> for ClassCounts {
-    fn add(&mut self, c: Classes) {
-        self.p[c as usize] += 1;
-    }
-
-    fn add_n(&mut self, c: Classes, n: usize) {
-        self.p[c as usize] += n;
-    }
-
-    fn probability(&self, c: Classes) -> f64 {
-        self.p[c as usize] as f64 / self.p.iter().sum::<usize>() as f64
-    }
-}
-
-impl<'a> Sum<&'a Classes> for ClassCounts {
-    fn sum<I: Iterator<Item=&'a Classes>>(iter: I) -> Self {
-        let mut counts = ClassCounts::new();
-        for c in iter {
-            counts.add(*c);
-        }
-        counts
-    }
-}
-
-impl IterMean<ClassCounts> for ClassCounts {
-    fn mean<I: ExactSizeIterator<Item=ClassCounts>>(iter: I) -> Self {
-        let mut total_counts = ClassCounts::new();
-        for c in iter {
-            for (a, b) in total_counts.p.iter_mut().zip(c.p.iter()) {
-                *a += *b;
-            }
-        }
-        total_counts
-    }
-}
+use common::rgb_classes::{ClassCounts, Classes};
 
 struct Sample<Y> {
     x: [f64; 2],
@@ -181,9 +117,9 @@ fn main() {
     // configure and fit random forest
     println!("Fitting...");
     let forest = DeterministicForestBuilder::new(
-        100,
+        100,  // 100 trees
         DeterministicTreeBuilder::new(
-            10,
+            10,  // don't split less than 10 samples
             BestRandomSplit::new(1)
         )
     ).fit(&mut data as &mut [_]);
