+++ examples/extra_trees_regressor.rs
@@ -2,6 +2,8 @@ extern crate forester;
 extern crate gnuplot;
 extern crate rand;
 
+mod common;
+
 use std::f64::consts::PI;
 
 use gnuplot::*;
@@ -11,21 +13,11 @@ use rand::{thread_rng, Rng};
 use forester::api::extra_trees_regressor::{ExtraTreesRegressor, Sample};
 use forester::vec2d::Vec2D;
 
+use common::numeric::{linspace, randspace};
+
 /// function used to generate training data
 fn func(x: &f64) -> f64 {
     (x * 2.0 * PI).sin() + thread_rng().gen::<f64>() * 0.2
-}
-
-/// generate a Vec<f64> of linearly spaced values
-fn linspace(l: f64, h: f64, n: usize) -> Vec<f64> {
-    let di = (h - l) / (n - 1) as f64;
-    (0..n).map(|i| l + di * i as f64).collect()
-}
-
-/// generate a Vec<f64> of uniformly distributed random values
-fn randspace(l: f64, h: f64, n: usize) -> Vec<f64> {
-    let mut rng = thread_rng();
-    (0..n).map(|_| rng.gen_range(l, h)).collect()
 }
 
 fn main() {
